requirements.txt
pandas
requests
html5lib
beautifulsoup4
pandas_gbq
google-cloud-bigquery